chore: move ds controllers to grid plugin
